@@ -1,7 +1,7 @@
 /*
   val3dity 
 
-  Copyright (c) 2011-2019, 3D geoinformation research group, TU Delft  
+  Copyright (c) 2011-2020, 3D geoinformation research group, TU Delft
 
   This file is part of val3dity.
 
@@ -31,12 +31,15 @@
 #include "Feature.h"
 #include "CityObject.h"
 #include "GenericObject.h"
+#include "IndoorModel.h"
+#include "IndoorGraph.h"
 #include "Surface.h"
 #include "MultiSurface.h"
 #include "CompositeSurface.h"
 #include "Solid.h"
 #include "CompositeSolid.h"
 #include "MultiSolid.h"
+#include "GeometryTemplate.h"
 
 
 using namespace std;
@@ -91,7 +94,7 @@ std::string IOErrors::get_report_text()
   {
     for (auto i : err.second)
     {
-      ss << err.first << " -- " << errorcode2description(err.first) << std::endl;
+      ss << err.first << " -- " << ALL_ERRORS[err.first] << std::endl;
       ss << "\tInfo: " << i << std::endl;
     }
   }
@@ -109,7 +112,7 @@ json IOErrors::get_report_json()
       json jj;
       jj["type"] = "Error";
       jj["code"] = err.first;
-      jj["description"] = errorcode2description(err.first);
+      jj["description"] = ALL_ERRORS[err.first];
       jj["info"] = i;
       j.push_back(jj);
     }
@@ -118,54 +121,16 @@ json IOErrors::get_report_json()
 }
 
 
-std::string errorcode2description(int code) {
-  switch(code)
-  {
-    case 0:   return string("STATUS_OK"); break;
-    //-- RING
-    case 101: return string("TOO_FEW_POINTS"); break;
-    case 102: return string("CONSECUTIVE_POINTS_SAME"); break;
-    case 103: return string("RING_NOT_CLOSED"); break;
-    case 104: return string("RING_SELF_INTERSECTION"); break;
-    //-- POLYGON
-    case 201: return string("INTERSECTION_RINGS"); break;
-    case 202: return string("DUPLICATED_RINGS"); break;
-    case 203: return string("NON_PLANAR_POLYGON_DISTANCE_PLANE"); break;
-    case 204: return string("NON_PLANAR_POLYGON_NORMALS_DEVIATION"); break;
-    case 205: return string("POLYGON_INTERIOR_DISCONNECTED"); break;
-    case 206: return string("INNER_RING_OUTSIDE"); break;
-    case 207: return string("INNER_RINGS_NESTED"); break;
-    case 208: return string("ORIENTATION_RINGS_SAME"); break;
-    //-- SHELL
-    case 300: return string("NOT_VALID_2_MANIFOLD"); break;
-    case 301: return string("TOO_FEW_POLYGONS"); break;
-    case 302: return string("SHELL_NOT_CLOSED"); break;
-    case 303: return string("NON_MANIFOLD_CASE"); break;
-    case 305: return string("MULTIPLE_CONNECTED_COMPONENTS"); break;
-    case 306: return string("SHELL_SELF_INTERSECTION"); break;
-    case 307: return string("POLYGON_WRONG_ORIENTATION"); break;
-    // case 309: return string("VERTICES_NOT_USED"); break;
-    //-- SOLID & MULTISOLID
-    case 401: return string("INTERSECTION_SHELLS"); break;
-    case 402: return string("DUPLICATED_SHELLS"); break;
-    case 403: return string("INNER_SHELL_OUTSIDE"); break;
-    case 404: return string("SOLID_INTERIOR_DISCONNECTED"); break;
-    case 405: return string("WRONG_ORIENTATION_SHELL"); break;
-    //-- COMPOSITESOLID
-    case 501: return string("INTERSECTION_SOLIDS"); break;
-    case 502: return string("DUPLICATED_SOLIDS"); break;
-    case 503: return string("DISCONNECTED_SOLIDS"); break;
-    //-- CityGML objects
-    case 601: return string("BUILDINGPARTS_OVERLAP"); break;
-    case 609: return string("CITYOBJECT_HAS_NO_GEOMETRY"); break;
-    //-- OTHERS
-    case 901: return string("INVALID_INPUT_FILE"); break;
-    case 902: return string("EMPTY_PRIMITIVE"); break;
-    case 903: return string("WRONG_INPUT_PARAMETERS"); break;
-    case 999: return string("UNKNOWN_ERROR"); break;
-    default:  return string("UNKNOWN_ERROR"); break;
-  }
+std::string IOErrors::get_input_file_type() {
+  return _inputfiletype;
 }
+
+
+void IOErrors::set_input_file_type(std::string s) {
+  _inputfiletype = s;
+}
+
+
 
 //-- ignore XML namespace
 std::string localise(std::string s)
@@ -443,42 +408,42 @@ void print_information(std::string &ifile)
   std::map<std::string, std::string> ns;
   pugi::xml_node ncm = doc.first_child();
   std::string vcitygml;
-  get_namespaces(ncm, ns, vcitygml);
+  get_namespaces(ncm, vcitygml);
   if (vcitygml.empty() == true) {
     std::cout << "File does not have the CityGML namespace. Abort." << std::endl;
     return;
   }
   std::cout << "++++++++++++++++++++ GENERAL +++++++++++++++++++++" << std::endl;
   std::cout << "CityGML version: " << vcitygml << std::endl;
-  report_primitives(doc, ns);
-  report_building(doc, ns);
+  report_primitives(doc);
+  report_building(doc);
 }
 
-void report_building(pugi::xml_document& doc, std::map<std::string, std::string>& ns) {
+void report_building(pugi::xml_document& doc) {
   std::cout << "++++++++++++++++++++ BUILDINGS +++++++++++++++++++" << std::endl;
   
-  std::string s = "//" + ns["building"] + "Building";
+  std::string s = "//" + NS["building"] + "Building";
   int nobuildings = doc.select_nodes(s.c_str()).size();
   print_info_aligned("Building", nobuildings);
 
-  s = "//" + ns["building"] + "Building" + "/" + ns["building"] + "consistsOfBuildingPart" + "[1]";
+  s = "//" + NS["building"] + "Building" + "/" + NS["building"] + "consistsOfBuildingPart" + "[1]";
   int nobwbp = doc.select_nodes(s.c_str()).size();
   print_info_aligned("without BuildingPart", (nobuildings - nobwbp), true);
   print_info_aligned("having BuildingPart", nobwbp, true);
-  s = "//" + ns["building"] + "Building" + "[@" + ns["gml"] + "id]";
+  s = "//" + NS["building"] + "Building" + "[@" + NS["gml"] + "id]";
   print_info_aligned("with gml:id", doc.select_nodes(s.c_str()).size(), true);
 
-  s = "//" + ns["building"] + "BuildingPart";
+  s = "//" + NS["building"] + "BuildingPart";
   int nobuildingparts = doc.select_nodes(s.c_str()).size();
   print_info_aligned("BuildingPart", nobuildingparts);
-  s = "//" + ns["building"] + "BuildingPart" + "[@" + ns["gml"] + "id]";
+  s = "//" + NS["building"] + "BuildingPart" + "[@" + NS["gml"] + "id]";
   print_info_aligned("with gml:id", doc.select_nodes(s.c_str()).size(), true);
   for (int lod = 1; lod <= 3; lod++) {
     std::cout << "LOD" << lod << std::endl;
     int totals = 0;
     int totalms = 0;
     int totalsem = 0;
-    report_building_each_lod(doc, ns, lod, totals, totalms, totalsem);
+    report_building_each_lod(doc, lod, totals, totalms, totalsem);
     print_info_aligned("Building stored in gml:Solid", totals, true);
     print_info_aligned("Building stored in gml:MultiSurface", totalms, true);
     print_info_aligned("Building with semantics for surfaces", totalsem, true);
@@ -494,15 +459,15 @@ void print_info_aligned(std::string o, size_t number, bool tab) {
   std::cout << std::setw(10) << std::right << number << std::endl;
 }
 
-void report_building_each_lod(pugi::xml_document& doc, std::map<std::string, std::string>& ns, int lod, int& total_solid, int& total_ms, int& total_sem) {
+void report_building_each_lod(pugi::xml_document& doc, int lod, int& total_solid, int& total_ms, int& total_sem) {
   total_solid = 0;
   total_ms = 0;
   total_sem = 0;
   std::string slod = "lod" + std::to_string(lod);
-  std::string s = "//" + ns["building"] + "Building";
+  std::string s = "//" + NS["building"] + "Building";
   pugi::xpath_node_set nb = doc.select_nodes(s.c_str());
   for (auto& b : nb) {
-    std::string s1 = ".//" + ns["building"] + slod + "Solid";
+    std::string s1 = ".//" + NS["building"] + slod + "Solid";
     pugi::xpath_node_set tmp = b.node().select_nodes(s1.c_str());
     if (tmp.empty() == false) {
       for (auto& nbp : tmp) {
@@ -510,7 +475,7 @@ void report_building_each_lod(pugi::xml_document& doc, std::map<std::string, std
         break;
       }
     }
-    s1 = ".//" + ns["building"] + slod + "MultiSurface";
+    s1 = ".//" + NS["building"] + slod + "MultiSurface";
     tmp = b.node().select_nodes(s1.c_str());
     if (tmp.empty() == false) {
       for (auto& nbp : tmp) {
@@ -518,7 +483,7 @@ void report_building_each_lod(pugi::xml_document& doc, std::map<std::string, std
         break;
       }
     }
-    s1 = ".//" + ns["building"] + "boundedBy" + "//" + ns["building"] + slod + "MultiSurface";
+    s1 = ".//" + NS["building"] + "boundedBy" + "//" + NS["building"] + slod + "MultiSurface";
     tmp = b.node().select_nodes(s1.c_str());
     if (tmp.empty() == false) {
       for (auto& nbp : tmp) {
@@ -571,25 +536,26 @@ void get_namespaces(pugi::xml_node& root, std::map<std::string, std::string>& ns
 }
 
 
-void report_primitives(pugi::xml_document& doc, std::map<std::string, std::string>& ns) {
+
+void report_primitives(pugi::xml_document& doc) {
   std::cout << "+++++++++++++++++++ PRIMITIVES +++++++++++++++++++" << std::endl;
   
-  std::string s = "//" + ns["gml"] + "Solid";
+  std::string s = "//" + NS["gml"] + "Solid";
   print_info_aligned("gml:Solid", doc.select_nodes(s.c_str()).size());
 
-  s = "//" + ns["gml"] + "MultiSolid";
+  s = "//" + NS["gml"] + "MultiSolid";
   print_info_aligned("gml:MultiSolid", doc.select_nodes(s.c_str()).size());
 
-  s = "//" + ns["gml"] + "CompositeSolid";
+  s = "//" + NS["gml"] + "CompositeSolid";
   print_info_aligned("gml:CompositeSolid", doc.select_nodes(s.c_str()).size());
   
-  s = "//" + ns["gml"] + "MultiSurface";
+  s = "//" + NS["gml"] + "MultiSurface";
   print_info_aligned("gml:MultiSurface", doc.select_nodes(s.c_str()).size());
   
-  s = "//" + ns["gml"] + "CompositeSurface";
+  s = "//" + NS["gml"] + "CompositeSurface";
   print_info_aligned("gml:CompositeSurface", doc.select_nodes(s.c_str()).size());
 
-  s = "//" + ns["gml"] + "Polygon";
+  s = "//" + NS["gml"] + "Polygon";
   print_info_aligned("gml:Polygon", doc.select_nodes(s.c_str()).size());
 
   std::cout << std::endl;
@@ -630,39 +596,65 @@ void process_json_surface(std::vector< std::vector<int> >& pgn, json& j, Surface
 }
 
 
-void read_file_cityjson(std::string &ifile, std::vector<Feature*>& lsFeatures, IOErrors& errs, double tol_snap)
+void process_json_geometries_of_co(json& jco, CityObject* co, std::vector<GeometryTemplate*>& lsGTs, json& j, double tol_snap)
 {
-  std::ifstream input(ifile);
-  json j;
-  try 
-  {
-    input >> j;
-    // TODO: other validation for CityJSON and just not JSON stuff?
-  }
-  catch (nlohmann::detail::parse_error e) 
-  {
-    errs.add_error(901, "Input file not a valid JSON file.");
-    return;
-  }
-  std::cout << "CityJSON input file" << std::endl;
-  std::cout << "# City Objects found: " << j["CityObjects"].size() << std::endl;
-  //-- compute (_minx, _miny)
-  compute_min_xy(j);
-  for (json::iterator it = j["CityObjects"].begin(); it != j["CityObjects"].end(); ++it) 
-  {
-    CityObject* co = new CityObject(it.key(), it.value()["type"]);
-    int idgeom = 0;
-    for (auto& g : it.value()["geometry"]) {
-      std::string theid = it.key() + "(" + std::to_string(idgeom) + ")";
-      if  (g["type"] == "Solid")
+  int idgeom = co->number_of_primitives();
+  for (auto& g : jco["geometry"]) {
+    std::string theid = co->get_id() + "(" + std::to_string(idgeom) + ")";
+    if  (g["type"] == "Solid")
+    {
+      Solid* s = new Solid(theid);
+      bool oshell = true;
+      int c = 0;
+      for (auto& shell : g["boundaries"]) 
       {
-        Solid* s = new Solid(theid);
-        bool oshell = true;
-        int c = 0;
-        for (auto& shell : g["boundaries"]) 
+        Surface* sh = new Surface(c, tol_snap);
+        c++;
+        for (auto& polygon : shell) { 
+          std::vector< std::vector<int> > pa = polygon;
+          process_json_surface(pa, j, sh);
+        }
+        if (oshell == true)
         {
-          Surface* sh = new Surface(c, tol_snap);
-          c++;
+          oshell = false;
+          s->set_oshell(sh);
+        }
+        else
+          s->add_ishell(sh);
+      }
+      co->add_primitive(s);
+    }
+    else if ( (g["type"] == "MultiSurface") || (g["type"] == "CompositeSurface") ) 
+    {
+      Surface* sh = new Surface(-1, tol_snap);
+      for (auto& p : g["boundaries"]) 
+      { 
+        std::vector< std::vector<int> > pa = p;
+        process_json_surface(pa, j, sh);
+      }
+      if (g["type"] == "MultiSurface")
+      {
+        MultiSurface* ms = new MultiSurface(theid);
+        ms->set_surface(sh);
+        co->add_primitive(ms);
+      }
+      else
+      {
+        CompositeSurface* cs = new CompositeSurface(theid);
+        cs->set_surface(sh);
+        co->add_primitive(cs);
+      }
+    }
+    else if (g["type"] == "MultiSolid") 
+    {
+      MultiSolid* ms = new MultiSolid(theid);
+      for (auto& solid : g["boundaries"]) 
+      {
+        Solid* s = new Solid();
+        bool oshell = true;
+        for (auto& shell : solid) 
+        {
+          Surface* sh = new Surface(-1, tol_snap);
           for (auto& polygon : shell) { 
             std::vector< std::vector<int> > pa = polygon;
             process_json_surface(pa, j, sh);
@@ -675,86 +667,327 @@ void read_file_cityjson(std::string &ifile, std::vector<Feature*>& lsFeatures, I
           else
             s->add_ishell(sh);
         }
-        co->add_primitive(s);
+        ms->add_solid(s);
       }
-      else if ( (g["type"] == "MultiSurface") || (g["type"] == "CompositeSurface") ) 
+      co->add_primitive(ms);
+    }
+    else if (g["type"] == "CompositeSolid") 
+    {
+      CompositeSolid* cs = new CompositeSolid(theid);
+      for (auto& solid : g["boundaries"]) 
       {
-        Surface* sh = new Surface(-1, tol_snap);
-        for (auto& p : g["boundaries"]) 
-        { 
-          std::vector< std::vector<int> > pa = p;
-          process_json_surface(pa, j, sh);
-        }
-        if (g["type"] == "MultiSurface")
+        Solid* s = new Solid();
+        bool oshell = true;
+        for (auto& shell : solid) 
         {
-          MultiSurface* ms = new MultiSurface(theid);
-          ms->set_surface(sh);
-          co->add_primitive(ms);
-        }
-        else
-        {
-          CompositeSurface* cs = new CompositeSurface(theid);
-          cs->set_surface(sh);
-          co->add_primitive(cs);
-        }
-      }
-      else if (g["type"] == "MultiSolid") 
-      {
-        MultiSolid* ms = new MultiSolid(theid);
-        for (auto& solid : g["boundaries"]) 
-        {
-          Solid* s = new Solid();
-          bool oshell = true;
-          for (auto& shell : solid) 
-          {
-            Surface* sh = new Surface(-1, tol_snap);
-            for (auto& polygon : shell) { 
-              std::vector< std::vector<int> > pa = polygon;
-              process_json_surface(pa, j, sh);
-            }
-            if (oshell == true)
-            {
-              oshell = false;
-              s->set_oshell(sh);
-            }
-            else
-              s->add_ishell(sh);
+          Surface* sh = new Surface(-1, tol_snap);
+          for (auto& polygon : shell) { 
+            std::vector< std::vector<int> > pa = polygon;
+            process_json_surface(pa, j, sh);
           }
-          ms->add_solid(s);
-        }
-        co->add_primitive(ms);
-      }
-      else if (g["type"] == "CompositeSolid") 
-      {
-        CompositeSolid* cs = new CompositeSolid(theid);
-        for (auto& solid : g["boundaries"]) 
-        {
-          Solid* s = new Solid();
-          bool oshell = true;
-          for (auto& shell : solid) 
+          if (oshell == true)
           {
-            Surface* sh = new Surface(-1, tol_snap);
-            for (auto& polygon : shell) { 
-              std::vector< std::vector<int> > pa = polygon;
-              process_json_surface(pa, j, sh);
-            }
-            if (oshell == true)
-            {
-              oshell = false;
-              s->set_oshell(sh);
-            }
-            else
-              s->add_ishell(sh);
+            oshell = false;
+            s->set_oshell(sh);
           }
-          cs->add_solid(s);
+          else
+            s->add_ishell(sh);
         }
-        co->add_primitive(cs);
-      }      
-      idgeom++;
+        cs->add_solid(s);
+      }
+      co->add_primitive(cs);
+    }
+    else if (g["type"] == "GeometryInstance") 
+    {
+      int gti = g["template"];
+      GeometryTemplate* g2 = lsGTs[gti];
+      co->add_primitive(g2);
+    }
+    idgeom++;
+  }
+}
+
+void read_file_cityjson(std::string &ifile, std::vector<Feature*>& lsFeatures, IOErrors& errs, double tol_snap)
+{
+  std::ifstream input(ifile);
+  json j;
+  try 
+  {
+    input >> j;
+  }
+  catch (nlohmann::detail::parse_error e) 
+  {
+    errs.add_error(901, "Input file not a valid JSON file.");
+    return;
+  }
+  // TODO: other validation for CityJSON or just let it crash?
+  if (j["type"] != "CityJSON") {
+    errs.add_error(901, "Input file not a CityJSON file.");
+    return;  
+  }
+  std::cout << "CityJSON input file" << std::endl;
+  std::cout << "# City Objects found: " << j["CityObjects"].size() << std::endl;
+  //-- compute (_minx, _miny)
+  compute_min_xy(j);
+  //-- read and store the GeometryTemplates
+  std::vector<GeometryTemplate*> lsGTs;
+  if (j.count("geometry-templates") == 1)
+  {
+    process_cityjson_geometrytemplates(j["geometry-templates"], lsGTs, tol_snap);
+  }
+  //-- process each CO
+  for (json::iterator it = j["CityObjects"].begin(); it != j["CityObjects"].end(); ++it) 
+  {
+    //-- BuildingParts geometries are put with those of a Building
+    if (it.value()["type"] == "BuildingPart")
+      continue;
+    CityObject* co = new CityObject(it.key(), it.value()["type"]);
+    process_json_geometries_of_co(it.value(), co, lsGTs, j, tol_snap);
+    //-- if Building has Parts, put them here in _lsPrimitives
+    if ( (it.value()["type"] == "Building") && (it.value().count("children") != 0) ) 
+    {
+      for (std::string bpid : it.value()["children"])
+      {
+        process_json_geometries_of_co(j["CityObjects"][bpid], co, lsGTs, j, tol_snap);
+      }
     }
     lsFeatures.push_back(co);
   }
 }
+
+
+void process_cityjson_geometrytemplates(json& j, std::vector<GeometryTemplate*>& lsGTs, double tol_snap)
+{
+  int count = 0;
+  for (auto& jt : j["templates"])
+  {
+    GeometryTemplate* gt = new GeometryTemplate(std::to_string(count));
+    if  (jt["type"] == "Solid")
+    {
+      Solid* s = new Solid("0");
+      bool oshell = true;
+      int c = 0;
+      for (auto& shell : jt["boundaries"]) 
+      {
+        Surface* sh = new Surface(c, tol_snap);
+        c++;
+        for (auto& polygon : shell) { 
+          std::vector< std::vector<int> > pa = polygon;
+          process_json_surface_geometrytemplate(pa, j, sh);
+        }
+        if (oshell == true)
+        {
+          oshell = false;
+          s->set_oshell(sh);
+        }
+        else
+          s->add_ishell(sh);
+      }
+      gt->add_primitive(s);
+    }
+    else if ( (jt["type"] == "MultiSurface") || (jt["type"] == "CompositeSurface") ) 
+    {
+      Surface* sh = new Surface(-1, tol_snap);
+      for (auto& p : jt["boundaries"]) 
+      { 
+        std::vector< std::vector<int> > pa = p;
+        process_json_surface_geometrytemplate(pa, j, sh);
+      }
+      if (jt["type"] == "MultiSurface")
+      {
+        MultiSurface* ms = new MultiSurface("0");
+        ms->set_surface(sh);
+        gt->add_primitive(ms);
+      }
+      else
+      {
+        CompositeSurface* cs = new CompositeSurface("0");
+        cs->set_surface(sh);
+        gt->add_primitive(cs);
+      }
+    }
+    lsGTs.push_back(gt);
+    count++;
+  }
+}
+
+
+void process_json_surface_geometrytemplate(std::vector< std::vector<int> >& pgn, json& j, Surface* sh)
+{
+  std::vector< std::vector<int> > pgnids;
+  for (auto& r : pgn)
+  {
+    std::vector<int> newr;
+    for (auto& i : r)
+    {
+      double x;
+      double y;
+      double z;
+      x = double(j["vertices-templates"][i][0]);
+      y = double(j["vertices-templates"][i][1]);
+      z = double(j["vertices-templates"][i][2]);
+      Point3 p3(x, y, z);
+      newr.push_back(sh->add_point(p3));
+    }
+    pgnids.push_back(newr);
+  }
+  sh->add_face(pgnids);
+}
+
+
+void process_gml_file_indoorgml(pugi::xml_document& doc, std::vector<Feature*>& lsFeatures, std::map<std::string, pugi::xpath_node>& dallpoly, IOErrors& errs, double tol_snap)
+{
+  //-- 0. read the header of the file and find its gml:name, if any
+  std::string nameim = "";
+  if (doc.first_child().attribute("gml:id") != 0) 
+    nameim = doc.first_child().attribute("gml:id").value();
+  else 
+    nameim = "MyIndoorModel";
+  IndoorModel* im = new IndoorModel(nameim);
+  lsFeatures.push_back(im);
+    
+  //-- 1. read each cellSpaceMember in the file (the primal objects)
+  //--    these can have different names, depending on the Extensions/ADEs used
+  std::string s = ".//" + NS["indoorgml"] + "cellSpaceMember";
+  pugi::xpath_node_set nn = doc.select_nodes(s.c_str());
+  int pcounter = 0;
+  for (pugi::xpath_node_set::const_iterator it = nn.begin(); it != nn.end(); ++it)
+  {
+    pugi::xml_node cs = it->node().first_child();
+    std::string theid   = "";
+    std::string duality = "";
+    //-- get the ID
+    if (cs.attribute("gml:id") != 0) {
+      // std::cout << "\t" << it->node().attribute("gml:id").value() << std::endl;
+      theid = cs.attribute("gml:id").value();
+    }
+    else 
+      theid = ("MISSING_ID_" + std::to_string(pcounter));
+    //-- get the duality pointer (max one, sweet)
+    s = NS["indoorgml"] + "duality";
+    for (pugi::xml_node child : cs.children(s.c_str()))
+    {
+      if (child.attribute("xlink:href") != 0) {
+        std::string s = child.attribute("xlink:href").value();
+        if (s[0] == '#')
+          s = s.substr(1);
+        duality = s;
+      }
+    }
+    // IndoorCell* cell = new IndoorCell(theid, duality);
+    //-- get the geometry, either Solid or Surface
+    s = NS["indoorgml"] + "cellSpaceGeometry";
+    Solid* sol;
+    for (pugi::xml_node child : cs.children(s.c_str()))
+    {
+      s = NS["indoorgml"] + "Geometry3D";
+      for (pugi::xml_node child2 : child.children(s.c_str()))
+      {
+        s = NS["gml"] + "Solid";
+        for (pugi::xml_node child3 : child2.children(s.c_str()))
+        {
+          // std::cout << "Solid: " << child3.attribute("gml:id").value() << std::endl;
+          sol = process_gml_solid(child3, dallpoly, tol_snap, errs);
+          if (sol->get_id() == "")
+            sol->set_id("MISSING_ID");
+          // cell->add_primitive(sol);
+        }
+      }
+    }
+    pcounter++;
+    im->add_cell(theid, sol, duality, cs.name());
+  }
+
+  //-- 2. read the dual graphs (yes there can be more than one) 
+  s = ".//" + NS["indoorgml"] + "SpaceLayer";
+  nn = doc.select_nodes(s.c_str());
+  for (pugi::xpath_node_set::const_iterator it = nn.begin(); it != nn.end(); ++it)
+  {
+    std::string idg;
+    if (it->node().attribute("gml:id") != 0) {
+      idg = it->node().attribute("gml:id").value();
+    }
+    else 
+      idg = "";
+    // IndoorGraph ig(idg);
+    IndoorGraph* ig = new IndoorGraph(idg);
+    //-- fetch all the edges
+    std::map<std::string, std::tuple<std::string,std::string>> edges;
+    s = ".//" + NS["indoorgml"] + "Transition";
+    pugi::xpath_node_set ntr = it->node().select_nodes(s.c_str());
+    for (pugi::xpath_node_set::const_iterator it = ntr.begin(); it != ntr.end(); ++it)
+    {
+      std::string theid = it->node().attribute("gml:id").value();
+      s = NS["indoorgml"] + "connects";
+      std::vector<std::string> connects;
+      for (pugi::xml_node child : it->node().children(s.c_str()))
+      {
+        if (child.attribute("xlink:href") != 0) {
+          std::string s = child.attribute("xlink:href").value();
+          if (s[0] == '#')
+            s = s.substr(1);
+          connects.push_back(s);
+        }
+      }
+      edges[theid] = std::make_tuple(connects[0], connects[1]);
+    }
+    //-- fetch all the nodes
+    s = ".//" + NS["indoorgml"] + "State";
+    pugi::xpath_node_set nstate = it->node().select_nodes(s.c_str());
+//    pugi::xpath_node_set nstate = doc.select_nodes(s.c_str());
+    for (pugi::xpath_node_set::const_iterator it = nstate.begin(); it != nstate.end(); ++it)
+    {
+      // std::cout << "---\n" << it->node().attribute("gml:id").value() << std::endl;
+      std::string vid = it->node().attribute("gml:id").value();
+      s = NS["indoorgml"] + "duality";
+      std::string vdual;
+      pugi::xml_node child = it->node().child(s.c_str());
+      if (child.attribute("xlink:href") != 0) {
+        vdual = child.attribute("xlink:href").value();
+        if (vdual[0] == '#')
+          vdual = vdual.substr(1);
+        // std::cout << "dual node: " << vdual << std::endl;
+      }
+      s = NS["indoorgml"] + "connects";
+      std::vector<std::string> vadj;
+      for (pugi::xml_node child : it->node().children(s.c_str()))
+      {
+        if (child.attribute("xlink:href") != 0) {
+          std::string s = child.attribute("xlink:href").value();
+          if (s[0] == '#')
+            s = s.substr(1);
+          if (std::get<1>(edges[s]) != vid)
+            vadj.push_back(std::get<1>(edges[s]));
+        }
+      }
+      s = ".//" + NS["gml"] + "pos";
+      pugi::xpath_node n = it->node().select_node(s.c_str());
+      // std::cout << n.node().child_value() << std::endl;
+      
+      std::string buf;
+      std::stringstream ss(n.node().child_value());
+      std::vector<std::string> tokens;
+      while (ss >> buf)
+        tokens.push_back(buf);
+      // long double x = std::stold(tokens[0]);
+      // long double y = std::stold(tokens[1]);
+      // long double z = std::stold(tokens[2]);
+      // TODO: use long double?
+      ig->add_vertex(vid, 
+                     std::stold(tokens[0]), 
+                     std::stold(tokens[1]), 
+                     std::stold(tokens[2]),
+                     vdual,
+                     vadj);
+    }
+    im->add_graph(ig);
+  }
+  if (im->get_number_graphs() == 0)
+    std::cout << "\t!!! No dual graph was found in the input file" << std::endl;
+}
+
+
+
 
 void process_gml_file_primitives(pugi::xml_document& doc, std::vector<Feature*>& lsFeatures, std::map<std::string, pugi::xpath_node>& dallpoly, IOErrors& errs, double tol_snap)
 {
@@ -942,6 +1175,13 @@ void read_file_gml(std::string &ifile, std::vector<Feature*>& lsFeatures, IOErro
   pugi::xml_node ncm = doc.first_child();
   std::string vcitygml;
   get_namespaces(ncm, vcitygml); //-- results in global variable NS in this unit
+
+  //-- CityGML v3 is not supported: warning to users
+  if (vcitygml == "v3.0") {
+    errs.add_error(904, "CityGML v3.0 files are not supported, use CityJSON (all versions fully supported) or downgrade to v2.0.");
+    return;
+  }
+
   if (NS.count("gml") == 0)
   {
     errs.add_error(901, "Input file does not have the GML namespace.");
@@ -955,7 +1195,13 @@ void read_file_gml(std::string &ifile, std::vector<Feature*>& lsFeatures, IOErro
   if ( (NS.count("citygml") != 0) && (ncm.name() == (NS["citygml"] + "CityModel")) )
   {
     std::cout << "CityGML input file" << std::endl;
+    errs.set_input_file_type("CityGML");
     process_gml_file_city_objects(doc, lsFeatures, dallpoly, errs, tol_snap, geom_is_sem_surfaces);
+  }
+  else if ( (NS.count("indoorgml") != 0) && (ncm.name() == (NS["indoorgml"] + "IndoorFeatures")) ) {
+    std::cout << "IndoorGML input file" << std::endl;
+    errs.set_input_file_type("IndoorGML");
+    process_gml_file_indoorgml(doc, lsFeatures, dallpoly, errs, tol_snap);
   }
   else
   {
@@ -985,9 +1231,17 @@ void get_namespaces(pugi::xml_node& root, std::string& vcitygml) {
         sns = "citygml";
         vcitygml = "v2.0";
       }
+      else if (value.find("http://www.opengis.net/citygml/3") != std::string::npos) {
+        sns = "citygml";
+        vcitygml = "v3.0";
+      }
       else if ( (value.find("http://www.opengis.net/gml") != std::string::npos) &&
                 (value.find("http://www.opengis.net/gmlcov") == std::string::npos) ) {
-          sns = "gml";
+        sns = "gml";
+      }
+      else if ( (value.find("http://www.opengis.net/indoorgml/1") != std::string::npos) &&
+                (value.find("core") != std::string::npos) ) {
+        sns = "indoorgml";
       }
       else if (value.find("http://www.opengis.net/citygml/building") != std::string::npos)
         sns = "building";
