@@ -47,7 +47,7 @@ master_doc = 'index'
 
 # General information about the project.
 project = u'val3dity'
-copyright = u'2011-2018, Hugo Ledoux'
+copyright = u'2011-2020, Hugo Ledoux'
 author = u'Hugo Ledoux'
 
 # The version info for the project you're documenting, acts as replacement for
@@ -55,9 +55,9 @@ author = u'Hugo Ledoux'
 # built documents.
 #
 # The short X.Y version.
-version = u'2.1'
+version = u'2.2.0'
 # The full version, including alpha/beta/rc tags.
-release = u'2.1'
+release = u'2.2.0'
 
 # The language for content autogenerated by Sphinx. Refer to documentation
 # for a list of supported languages.
