@@ -32,6 +32,7 @@
 #include "input.h"
 #include "validate_shell.h"
 #include <CGAL/Polygon_mesh_processing/self_intersections.h>
+#include <CGAL/Side_of_triangle_mesh.h>
 #include <geos_c.h>
 #include <sstream>
 
@@ -84,7 +85,7 @@ void Surface::add_error(int code, std::string faceid, std::string info)
 {
   std::tuple<std::string, std::string> a(faceid, info);
   _errors[code].push_back(a);
-  std::clog << "\tERROR " << code << ": " << errorcode2description(code);
+  std::clog << "\tERROR " << code << ": " << ALL_ERRORS[code];
   if (faceid.empty() == false)
     std::clog << " (face " << faceid << ")";
   std::clog << std::endl;
@@ -113,7 +114,7 @@ json Surface::get_report_json()
       json jj;
       jj["type"] = "Error";
       jj["code"] = std::get<0>(err);
-      jj["description"] = errorcode2description(std::get<0>(err));
+      jj["description"] = ALL_ERRORS[std::get<0>(err)];
       jj["id"] = std::get<0>(e);
       jj["info"] = std::get<1>(e);
       j.push_back(jj);
@@ -913,5 +914,24 @@ bool Surface::has_face_rings_toofewpoints(const std::vector< std::vector<int> >&
   }
   return bErrors;
 }
+
+int Surface::side_of_triangle_surface(Point3& p)
+  /*
+   -2 = not valid polyhedron
+   -1 = outside
+   0 = directly on the boundary of polyhedron
+   1 = inside
+   */
+{
+  int re = -2;
+  if ( (_polyhedron != NULL) && (CGAL::is_triangle_mesh(*_polyhedron) == true) )
+  {
+    CGAL::Side_of_triangle_mesh<CgalPolyhedron, K> inside(*_polyhedron);
+    Point3 p_translated(p.x() - Surface::_shiftx, p.y() - Surface::_shifty, p.z());
+    re = inside(p_translated);
+  }
+  return re;
+}
+
 
 } // namespace val3dity
