@@ -95,12 +95,13 @@ typedef long long int64;
 
 typedef enum
 {
-  SOLID            = 0,
-  COMPOSITESOLID   = 1,
-  MULTISOLID       = 2,
-  COMPOSITESURFACE = 3,
-  MULTISURFACE     = 4,
-  ALL              = 5,
+  SOLID             = 0,
+  COMPOSITESOLID    = 1,
+  MULTISOLID        = 2,
+  COMPOSITESURFACE  = 3,
+  MULTISURFACE      = 4,
+  GEOMETRYTEMPLATE  = 5,
+  ALL               = 9,
 } Primitive3D;
 
 typedef enum
@@ -118,6 +119,56 @@ struct Error {
   int         errorcode;
   std::string info1;
   std::string info2;
+};
+
+static std::map<int,std::string> ALL_ERRORS = {
+  //-- RING
+  {101, "TOO_FEW_POINTS"}, 
+  {102, "CONSECUTIVE_POINTS_SAME"}, 
+  {103, "RING_NOT_CLOSED"}, 
+  {104, "RING_SELF_INTERSECTION"}, 
+  //-- POLYGON
+  {201, "INTERSECTION_RINGS"}, 
+  {202, "DUPLICATED_RINGS"}, 
+  {203, "NON_PLANAR_POLYGON_DISTANCE_PLANE"}, 
+  {204, "NON_PLANAR_POLYGON_NORMALS_DEVIATION"}, 
+  {205, "POLYGON_INTERIOR_DISCONNECTED"}, 
+  {206, "INNER_RING_OUTSIDE"}, 
+  {207, "INNER_RINGS_NESTED"}, 
+  {208, "ORIENTATION_RINGS_SAME"}, 
+  //-- SHELL
+  {300, "NOT_VALID_2_MANIFOLD"}, 
+  {301, "TOO_FEW_POLYGONS"}, 
+  {302, "SHELL_NOT_CLOSED"}, 
+  {303, "NON_MANIFOLD_CASE"}, 
+  {305, "MULTIPLE_CONNECTED_COMPONENTS"}, 
+  {306, "SHELL_SELF_INTERSECTION"}, 
+  {307, "POLYGON_WRONG_ORIENTATION"}, 
+  // {309, "VERTICES_NOT_USED"}, 
+  //-- SOLID & MULTISOLID
+  {401, "INTERSECTION_SHELLS"}, 
+  {402, "DUPLICATED_SHELLS"}, 
+  {403, "INNER_SHELL_OUTSIDE"}, 
+  {404, "SOLID_INTERIOR_DISCONNECTED"}, 
+  {405, "WRONG_ORIENTATION_SHELL"}, 
+  //-- COMPOSITESOLID
+  {501, "INTERSECTION_SOLIDS"}, 
+  {502, "DUPLICATED_SOLIDS"}, 
+  {503, "DISCONNECTED_SOLIDS"}, 
+  //-- CityGML objects
+  {601, "BUILDINGPARTS_OVERLAP"}, 
+  {609, "CITYOBJECT_HAS_NO_GEOMETRY"}, 
+  //-- IndoorGML objects
+  {701, "CELLS_OVERLAP"}, 
+  {702, "DUAL_VERTEX_OUTSIDE_CELL"}, 
+  {703, "PRIMAL_DUAL_XLINKS_ERROR"}, 
+  {704, "PRIMAL_DUAL_ADJACENCIES_INCONSISTENT"}, 
+  //-- OTHERS
+  {901, "INVALID_INPUT_FILE"}, 
+  {902, "EMPTY_PRIMITIVE"}, 
+  {903, "WRONG_INPUT_PARAMETERS"}, 
+  {904, "FORMAT_NOT_SUPPORTED"}, 
+  {999, "UNKNOWN_ERROR"} 
 };
 
 } // namespace val3dity
